% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.function.R
\name{testing.function}
\alias{testing.function}
\title{Testing function}
\usage{
testing.function(test.word)
}
\description{
This is a testing function
}
\examples{
testing.function()
}
\keyword{none}

